% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theoretical_gwas.R
\name{add_ld_to_params}
\alias{add_ld_to_params}
\title{Modify SNP effects to account for LD}
\usage{
add_ld_to_params(
  params,
  ldobj = NULL,
  ldobjlist = NULL,
  ldobjfiles = NULL,
  nthreads = 1
)
}
\arguments{
\item{params}{Output from \code{generate_gwas_params}}

\item{ldobj}{LD objects (e.g. see test_ldobj)}

\item{ldobjlist}{List of LD objects}

\item{ldobjfiles}{Array of filenames containing LD object files (e.g. see \code{generate_ldobj})}

\item{nthreads}{Number of threads (can be slow for complete GWAS and large LD regions)}
}
\value{
Updated params
}
\description{
After generating the set of causal effects at each SNP, use an LD correlation matrix to transform the effects to reflect the correlation structure at the SNPs. Note if running many repeats, only need to generate the LD-modified params once and then can repeatedly re-sample using generate_gwas_ss
}
