% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theoretical_gwas.R
\name{arbitrary_map}
\alias{arbitrary_map}
\title{Create an arbitrary map}
\usage{
arbitrary_map(af)
}
\arguments{
\item{af}{vectory of allele frequencies}
}
\value{
data frame with snp name, chr, pos, ea, oa
}
\description{
Create an arbitrary map
}
