% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.r
\name{ascertain_samples}
\alias{ascertain_samples}
\title{Ascertain some proportion of cases and controls from binary phenotype}
\usage{
ascertain_samples(d, prop_cases)
}
\arguments{
\item{d}{Vector of 1/0}

\item{prop_cases}{Proportion of 1s to retain}
}
\value{
Array of IDs
}
\description{
Ascertain some proportion of cases and controls from binary phenotype
}
