% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system.r
\name{create_system}
\alias{create_system}
\title{Wrapper for simulation pipeline}
\usage{
create_system(
  nidx,
  nidy,
  nidu = 0,
  nu = 0,
  na = 0,
  nb = 0,
  var_x.y,
  nsnp_x,
  var_gx.x,
  var_gx.y = 0,
  mu_gx.y = 0,
  prop_gx.y = 1,
  nsnp_y = 0,
  var_gy.y = 0,
  var_gy.x = 0,
  mu_gy.x = 0,
  prop_gy.x = 1
)
}
\arguments{
\item{nidx}{}

\item{nidy}{}

\item{var_x.y}{}

\item{nsnp_x}{}

\item{var_gx.x}{}

\item{nidu=0}{If 0 then don't simulate separate populations for the u variables}

\item{nu=0}{Number of confounders influencing x and y}

\item{na=0}{Number of traits upstream of x}

\item{nb=0}{Number of traits upstream of y}

\item{var_gx.y=0}{}

\item{mu_gx.y=0}{}

\item{prop_gx.y=1}{}

\item{nsnp_y=0}{}

\item{var_gy.y=0}{}

\item{var_gy.x=0}{}

\item{mu_gy.x=0}{}

\item{prop_gy.x=1}{}
}
\value{
List of effects across system
}
\description{
Based on the parameters specified this function will call \code{init_parameters}, \code{add_u}, \code{sample_system_effects}, \code{simulate_population} and \code{estimate_system_effects}. A separate population is generated for each phenotype (x, y and each of u) to allow 2SMR and PRS analyses
}
