% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr.R
\name{dat_sign}
\alias{dat_sign}
\title{Flip the signs in dat to positive exposure values}
\usage{
dat_sign(dat)
}
\arguments{
\item{dat}{Output from get_effs or harmonise_data}
}
\value{
data frame
}
\description{
Flip the signs in dat to positive exposure values
}
