% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system.r
\name{estimate_system_effects}
\alias{estimate_system_effects}
\title{Estimate the effects of all SNPs on all phenotypes}
\usage{
estimate_system_effects(sim)
}
\arguments{
\item{sim}{Output from \code{simulate_population}}
}
\value{
Lists of SNP-trait effect estimates
}
\description{
Estimate the effects of all SNPs on all phenotypes
}
