% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.r
\name{expected_gwas}
\alias{expected_gwas}
\title{For known effect size, sample size and variances, what is the expected association?}
\usage{
expected_gwas(eff, n, vx, vy)
}
\arguments{
\item{eff}{Vector of effect sizes}

\item{n}{Vector of sample sizes}

\item{vx}{Vector of variances of x}

\item{vy}{vector of varinces of y}
}
\value{
Data frame of associations
}
\description{
For known effect size, sample size and variances, what is the expected association?
}

