% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theoretical_gwas.R
\name{expected_mse}
\alias{expected_mse}
\title{Calculate expected MSE}
\usage{
expected_mse(beta, maf, vy)
}
\arguments{
\item{beta}{<what param does>}

\item{maf}{<what param does>}

\item{vy}{<what param does>}
}
\value{
Numeric
}
\description{
Calculate expected MSE
}
