% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theoretical_gwas.R
\name{expected_se}
\alias{expected_se}
\title{Expected se given beta, af, n and vy}
\usage{
expected_se(beta, af, n, vy)
}
\arguments{
\item{beta}{array of effect sizes}

\item{af}{array of allele frequencies}

\item{n}{sample size}

\item{vy}{variance of y}
}
\value{
array of standard errors
}
\description{
se = sqrt(sigma_e^2 / ss(x))
}
