% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theoretical_gwas.R
\name{expected_ssx}
\alias{expected_ssx}
\title{Calculate expected SSX}
\usage{
expected_ssx(af, n)
}
\arguments{
\item{af}{array of allele frequencies}

\item{n}{sample size}
}
\value{
Numeric
}
\description{
Calculate expected SSX
}
