% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.r, R/mr.R
\name{fast_assoc}
\alias{fast_assoc}
\alias{fast_assoc}
\title{Get summary statistics in simple linear regression}
\usage{
fast_assoc(y, x)

fast_assoc(y, x)
}
\arguments{
\item{y}{Vector of dependent variable}

\item{x}{Vector of independent variable}

\item{y}{Vector (outcome)}

\item{x}{Vector (exposure)}
}
\value{
List

list of linear regression results
}
\description{
Get summary statistics in simple linear regression

Quickly perform linear regression of one variable on another
}
