% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theoretical_gwas.R
\name{generate_gwas_params}
\alias{generate_gwas_params}
\title{Generate SNP effects given MAF, h2 and selection}
\usage{
generate_gwas_params(maf, h2, S = 0)
}
\arguments{
\item{maf}{Vector of allele frequencies, one for each SNP}

\item{h2}{Variance explained by all SNPs}

\item{S}{Selection coefficient on trait. Default = 0}
}
\value{
data frame of maf and beta
}
\description{
Generate SNP effects given MAF, h2 and selection
}
