% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ld.r
\name{generate_ldobj}
\alias{generate_ldobj}
\title{<brief desc>}
\usage{
generate_ldobj(
  outdir,
  bfile,
  regions,
  plink_bin = genetics.binaRies::get_plink_binary(),
  nthreads = 1
)
}
\arguments{
\item{bfile}{<what param does>}

\item{regionfile}{<what param does>}

\item{plink_bin=genetics.binaRies::get_plink_binary()}{<what param does>}

\item{nthreads=1}{<what param does>}
}
\value{

}
\description{
<full description>
}
