% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.r, R/mr.R
\name{get_effs}
\alias{get_effs}
\alias{get_effs}
\title{Get effs for two traits and make dat format}
\usage{
get_effs(x, y, g, en = "X", on = "Y")

get_effs(x, y, g, en = "X", on = "Y")
}
\arguments{
\item{x}{Vector of exposure phenotype}

\item{y}{Vector of outcome phenotype}

\item{g}{Matrix of genotypes}

\item{x}{Vector of exposure phenotype}

\item{y}{Vector of outcome phenotype}

\item{g}{Matrix of genotypes from make_geno}

\item{en="X"}{Name of exposure}

\item{on="Y"}{Name of outcome}
}
\value{
Data frame

Data frame of dat format
}
\description{
Get effs for two traits and make dat format

Provide the exposure trait and the outcome trait plus the genotype matrix, and this function will perform GWAS on each and produce dat format data frame
}
