% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ld.r
\name{get_ld}
\alias{get_ld}
\title{Get LD matrix for a specified region from bfile reference panel}
\usage{
get_ld(chr, from, to, bfile, plink_bin = genetics.binaRies::get_plink_binary())
}
\arguments{
\item{chr}{Chromosome}

\item{from}{from bp}

\item{to}{to bp}

\item{bfile}{LD reference panel}

\item{plink_bin}{Plink binary default=genetics.binaRies::get_plink_binary()}
}
\value{
List of LD matrix and map info
}
\description{
Get LD matrix for a specified region from bfile reference panel
}
