% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ld.r
\name{get_regions_from_ldobjdir}
\alias{get_regions_from_ldobjdir}
\title{Determine regions from LD file}
\usage{
get_regions_from_ldobjdir(ldobjdir)
}
\arguments{
\item{ldobjdir}{Directory containing output from \code{generate_ldobj}}
}
\value{
Data frame
}
\description{
Determine regions from LD file
}
