% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.r, R/mr.R
\name{gwas}
\alias{gwas}
\alias{gwas}
\title{Perform association of many SNPs against phenotype}
\usage{
gwas(y, g)

gwas(y, g)
}
\arguments{
\item{y}{Vector of phenotypes}

\item{g}{Matrix of genotypes}

\item{y}{vector}

\item{g}{Matrix (output from make_geno)}
}
\value{
Data frame

Data frame of results
}
\description{
Perform association of many SNPs against phenotype

Quickly perform GWAS of y against genotype matrix
}
