% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theoretical_gwas.R
\name{gwas_se}
\alias{gwas_se}
\title{Get the expected se for a gwas given n, h2, beta maf}
\usage{
gwas_se(n, h2, beta, maf)
}
\arguments{
\item{n}{sample size}

\item{h2}{heritability}

\item{beta}{array of effect sizes}

\item{maf}{array of allele frequencies}
}
\value{
array of standard errors
}
\description{
Get the expected se for a gwas given n, h2, beta maf
}
