% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk.r
\name{gx_to_gp}
\alias{gx_to_gp}
\title{Translate risk from liability to probability scale}
\usage{
gx_to_gp(gx, h2x, prev)
}
\arguments{
\item{gx}{score on liability scale}

\item{h2x}{Disease heritability on liability scale}

\item{prev}{Prevalence}
}
\value{
Vector of disease probabilities
}
\description{
Translate risk from liability to probability scale
}
