% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ld.r
\name{hap_freqs}
\alias{hap_freqs}
\title{Estimate haplotype frequencies for two loci}
\usage{
hap_freqs(r, p1, p2)
}
\arguments{
\item{r}{Required LD r}

\item{p1}{Freq 1}

\item{p2}{Freq 2}
}
\value{
vector
}
\description{
Estimate haplotype frequencies for two loci
}
