% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldetect.r
\docType{data}
\name{ldetect}
\alias{ldetect}
\title{Data frame of independent LD regions}
\format{
A data frame with 5731 rows and 4 columns
\describe{
\item{chr}{chromosome}
\item{start}{start position}
\item{stop}{stop position}
\item{pop}{Population}
}
}
\source{
\url{https://bitbucket.org/nygcresearch/ldetect-data](https://bitbucket.org/nygcresearch/ldetect-data/src/master/}
}
\usage{
ldetect
}
\description{
Taken from Berisa and Pickrell (2016) - \url{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC4731402/} - a list of independent LD regions for Africans, Europeans and Asians. All in build hg19
}
\keyword{datasets}
