% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsq_liability.r
\name{lor_to_rsq}
\alias{lor_to_rsq}
\title{Estimate proportion of variance of liability explained by SNP in general population}
\usage{
lor_to_rsq(b, af, ncase, ncontrol, prevalence, model = "logit")
}
\arguments{
\item{b}{Log odds ratio}

\item{af}{}

\item{ncase}{<what param does>}

\item{ncontrol}{<what param does>}

\item{prevalence}{<what param does>}

\item{model}{Is the effect size estiamted in "logit" (default) or "probit" model}
}
\value{
Rsq
}
\description{
This uses equation 10 in Genetic Epidemiology 36 : 214–224 (2012)
}
