% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.r
\name{make_dat}
\alias{make_dat}
\title{Organise outputs from \code{gwas} into harmonised dat format}
\usage{
make_dat(gwasx, gwasy)
}
\arguments{
\item{gwasx}{Output from \code{gwas}}

\item{gwasy}{Output from \code{gwas}}
}
\description{
Organise outputs from \code{gwas} into harmonised dat format
}
