% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.r
\name{make_geno}
\alias{make_geno}
\title{Create genotype matrix}
\usage{
make_geno(nid, nsnp, af)
}
\arguments{
\item{nid}{Number of samples}

\item{nsnp}{Number of SNPs}

\item{af}{Allele frequency of all SNPs (all SNPs have the same allele frequency)}
}
\value{
Matrix of genotypes, rows = individuals, columns = snps
}
\description{
Create genotype matrix
}

