% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr.R
\name{make_mvdat}
\alias{make_mvdat}
\title{Take several exposures and one outcome and make the data required for multivariable MR}
\usage{
make_mvdat(exposures, y, g)
}
\arguments{
\item{exposures}{List of exposure vectors}

\item{y}{Vector of outcomes}

\item{g}{Matrix of genotypes}
}
\value{
mv_harmonise_data output
}
\description{
Take several exposures and one outcome and make the data required for multivariable MR
}
