% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.r
\name{make_phen}
\alias{make_phen}
\title{Create phenotype from vector of effects and matrix of factors}
\usage{
make_phen(effs, indep, vy = 1, vx = rep(1, length(effs)), my = 0)
}
\arguments{
\item{effs}{Vector of effects. If vy and vx are set to default then effect sizes are equivalent to sqrt(variance explained) of each factor}

\item{indep}{Matrix of factors. Must be same number of columns as length of effs}

\item{vy=1}{Variance of phenotype to be outputed}

\item{vx=rep(1, }{length(effs)) Vector of variances of each factor in indep}

\item{my=0}{Mean of phenotype to be outputted}
}
\value{
Phenotype vector
}
\description{
Create phenotype from vector of effects and matrix of factors
}

