% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_mr.r
\name{merge_exp_out}
\alias{merge_exp_out}
\title{Organise outputs from \code{gwas} into harmonised dat format}
\usage{
merge_exp_out(gwasx, gwasy, xname = "X", yname = "Y")
}
\arguments{
\item{gwasx}{Output from \code{gwas}}

\item{gwasy}{Output from \code{gwas}}

\item{xname}{exposure name}

\item{yname}{outcome name}
}
\value{
data frame
}
\description{
Organise outputs from \code{gwas} into harmonised dat format
}
