% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk.r
\name{range01}
\alias{range01}
\title{Scale variable to have range between 0 and 1}
\usage{
range01(x)
}
\arguments{
\item{x}{Vector}
}
\value{
Vector
}
\description{
Scale variable to have range between 0 and 1
}

