% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ld.r
\name{read_ldobjdir}
\alias{read_ldobjdir}
\title{Read in LD objects into list}
\usage{
read_ldobjdir(ldobjdir, nthreads = 1)
}
\arguments{
\item{ldobjdir}{Directory containing output from \code{generate_ldobj}}

\item{nthreads}{Number of threads. Default=1}
}
\value{
List of ldobj
}
\description{
Read in LD objects into list
}
