% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr.R
\name{recode_dat}
\alias{recode_dat}
\alias{recode_dat}
\title{Recode the summary data so that everything is relative to positive exposure effect sizes}
\usage{
recode_dat(dat)

recode_dat(dat)
}
\arguments{
\item{dat}{Output from harmonise_data}

\item{dat}{Output from get_effs or harmonise_data}
}
\value{
Data frame

data frame
}
\description{
Recode the summary data so that everything is relative to positive exposure effect sizes

Simular to dat_sign but flips around the regression line, which does not go through the intercept
}
