% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.r
\name{recode_dat_intercept}
\alias{recode_dat_intercept}
\title{Intercept recoding to have every effect on x positive}
\usage{
recode_dat_intercept(dat)
}
\arguments{
\item{dat}{Output from get_effs}
}
\value{
Data frame
}
\description{
Tries to avoid issue of recoding by finding intercept and pivoting negative g-x associations around intercept
}

