% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_mr.r
\name{recode_dat_simple}
\alias{recode_dat_simple}
\title{Simple recoding to have every effect on x positive}
\usage{
recode_dat_simple(dat)
}
\arguments{
\item{dat}{Output from get_effs}
}
\value{
Data frame
}
\description{
Simple recoding to have every effect on x positive
}
