% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk.r
\name{risk_cross_plot}
\alias{risk_cross_plot}
\title{Plot liability vs probability disease risk}
\usage{
risk_cross_plot(
  x,
  o,
  xlab = "Values (low to high)",
  ylab = "",
  title = "",
  xname = "GRS",
  oname = "Disease"
)
}
\arguments{
\item{x}{Disease risk on liability scale}

\item{o}{Disease risk on probability scale}

\item{xlab}{default="Values (low to high)"}

\item{ylab}{default=""}

\item{title}{default=""}

\item{xname}{Name of liability. Default="GRS"}

\item{oname}{Name of disease. Default="Disease"}
}
\value{
ggplot
}
\description{
Plot liability vs probability disease risk
}
