% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk.r
\name{risk_simulation}
\alias{risk_simulation}
\title{Make simulation to compare disease and liability scales}
\usage{
risk_simulation(G, eff, prevalence, prop_discovered)
}
\arguments{
\item{G}{Matrix of genotypes}

\item{eff}{SNP effects on liability scale}

\item{prevalence}{Disease prevalence}

\item{prop_discovered}{Proportion of SNPs discovered}
}
\value{
Data frame
}
\description{
Compares the liability and probability of disease under two scenarios - where all SNPs are known, and where only some SNPs are known
}

