% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system.r
\name{sample_system_effects}
\alias{sample_system_effects}
\title{Sample the actual effects based on initial parameters}
\usage{
sample_system_effects(parameters)
}
\arguments{
\item{parameters}{Output from \code{init_parameters} or \code{add_u}}
}
\value{
List of effect parameters
}
\description{
Sample the actual effects based on initial parameters
}
