% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ld.r
\name{simulate_geno}
\alias{simulate_geno}
\title{Simulate genotypes from haplotypes}
\usage{
simulate_geno(nid, r, p1, p2)
}
\arguments{
\item{nid}{Number of samples}

\item{r}{Desired LD r}

\item{p1}{Freq 1}

\item{p2}{Freq 2}
}
\value{
Matrix
}
\description{
Simulate genotypes from haplotypes
}
