% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ld.r
\name{simulate_haplotypes}
\alias{simulate_haplotypes}
\title{Simulate haplotypes of two loci}
\usage{
simulate_haplotypes(nid, r, p1, p2)
}
\arguments{
\item{nid}{Number of samples}

\item{r}{Desired LD r}

\item{p1}{Freq 1}

\item{p2}{Freq 2}
}
\value{
Matrix
}
\description{
Simulate haplotypes of two loci
}
