% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system.r
\name{simulate_population}
\alias{simulate_population}
\title{Simulate individual level data from initial parameters}
\usage{
simulate_population(parameters, nid)
}
\arguments{
\item{parameters}{Output from \code{init_parameters} or \code{add_u}}

\item{nid}{Sample size to generate}
}
\value{
List of matrices and vectors that represent individual level data
}
\description{
Simulate individual level data from initial parameters
}
