% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_set.r
\name{summary_set}
\alias{summary_set}
\title{Wrapper for generating a summary set}
\usage{
summary_set(
  beta_gx,
  beta_gy,
  af,
  n_gx,
  n_gy,
  n_overlap,
  cor_xy,
  prev_y = NA,
  sigma_x = 1,
  sigma_y = 1
)
}
\arguments{
\item{beta_gx}{Array of true effects on x}

\item{beta_gy}{Array of true effects on y}

\item{af}{Array of effect allele frequencies}

\item{n_gx}{Sample size of g-x association}

\item{n_gy}{Sample size of g-y association}

\item{n_overlap}{Number of overlapping samples}

\item{cor_xy}{Observational correlation between x and y}

\item{prev_y}{Disease prevalence of y. Default = NA (in which case treated as continuous)}

\item{sigma_x}{SD of x. Default=1}

\item{sigma_y}{SD of y. Default=1}
}
\value{
Data frame of summary statistics for x and y
}
\description{
Allows arbitrary sample overlap
}
