% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ld.r
\name{test_ldobj}
\alias{test_ldobj}
\title{Create test LD object}
\usage{
test_ldobj(nsnp, chunksize)
}
\arguments{
\item{nsnp}{Number of SNPs}

\item{chunksize}{Chunksize for splitting}
}
\value{
list of chunks, which each contain map and LD matrix
}
\description{
Create test LD object
}
