% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system.r
\name{test_system}
\alias{test_system}
\title{Apply MR tests to system}
\usage{
test_system(ss, id = "test")
}
\arguments{
\item{ss}{Output from create_syste,}

\item{id}{string denoting simulation ID}
}
\value{
List
}
\description{
Apply MR tests to system
}
