% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theoretical_gwas.R
\name{theoretical_gwas}
\alias{theoretical_gwas}
\title{Create a theoretical GWAS dataset}
\usage{
theoretical_gwas(beta, maf, h2, nid, minmaf = 0.01)
}
\arguments{
\item{beta}{Array of beta values}

\item{maf}{array of maf values}

\item{h2}{h2 of trait}

\item{nid}{sample size}

\item{minmaf}{minimum allowed maf. default=0.01 to prevent instability}
}
\value{
list of data frames
}
\description{
Choose nsnp, nid, h2
Create true effects
Create expected se
Create sampled effects
}
