% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr_system.r
\name{add_u}
\alias{add_u}
\title{Add confounder variables and their instruments}
\usage{
add_u(parameters, nsnp_u, var_u.x, var_u.y, var_gu.u)
}
\arguments{
\item{parameters}{Output from \code{init_parameters}}

\item{nsnp_u}{nsnp_u}

\item{var_u.x}{var_u.x}

\item{var_u.y}{var_u.y}

\item{var_gu.u}{var_gu.u}
}
\value{
List of model parameters
}
\description{
Add confounder variables and their instruments
}
