% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ld.r
\name{correlated_binomial}
\alias{correlated_binomial}
\title{Simulate two correlated binomial variables}
\usage{
correlated_binomial(nid, p1, p2, rho, n = 2, round = TRUE, print = FALSE)
}
\arguments{
\item{nid}{Number of samples}

\item{p1}{Frequency 1}

\item{p2}{Frequency 2}

\item{rho}{Target correlation}

\item{n}{Binomial parameter, should be 2 (default) for genotypes}

\item{round}{Round or not Default=TRUE}

\item{print}{Print or not Default=FALSE}
}
\value{
Matrix
}
\description{
Simulate two correlated binomial variables
}
