% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theoretical_gwas.R
\name{expected_mse}
\alias{expected_mse}
\title{Calculate expected MSE}
\usage{
expected_mse(beta, af, vy)
}
\arguments{
\item{beta}{array of effect sizes}

\item{af}{array of allele frequencies}

\item{vy}{variance of y}
}
\value{
Numeric
}
\description{
Calculate expected MSE
}
