% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theoretical_gwas.R
\name{expected_ssx}
\alias{expected_ssx}
\title{Calculate expected SSX}
\usage{
expected_ssx(maf, n)
}
\arguments{
\item{maf}{<what param does>}

\item{n}{<what param does>}
}
\value{
Numeric
}
\description{
Calculate expected SSX
}
