% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.r
\name{fast_assoc}
\alias{fast_assoc}
\title{Get summary statistics in simple linear regression}
\usage{
fast_assoc(y, x)
}
\arguments{
\item{y}{Vector of dependent variable}

\item{x}{Vector of independent variable}
}
\value{
List
}
\description{
Get summary statistics in simple linear regression
}

