% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theoretical_gwas.R
\name{generate_gwas_params}
\alias{generate_gwas_params}
\title{Generate SNP effects given MAF, h2 and selection}
\usage{
generate_gwas_params(af, h2, S = 0, Pi = 1, snp = NULL)
}
\arguments{
\item{af}{Vector of effect allele frequencies, one for each SNP}

\item{h2}{Variance explained by all SNPs}

\item{S}{Selection coefficient on trait. Default = 0}

\item{Pi}{Proportion of variants that have an effect - sampled randomly. Default=1}

\item{snp}{Variant identifiers. Default NULL means 1:nsnp}
}
\value{
data frame
}
\description{
Generate SNP effects given MAF, h2 and selection
}
