% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theoretical_gwas.R
\name{generate_gwas_params}
\alias{generate_gwas_params}
\title{Generate SNP effects given MAF, h2 and selection}
\usage{
generate_gwas_params(map, h2, S = 0, Pi = 1)
}
\arguments{
\item{map}{Data frame containing at least `af` allele frequency and `snp` SNP columns. SNPs must be unique, `af` must be between 0 and 1. Optionally also include the chr, pos, ref, alt columns if using LD-aware simulations}

\item{h2}{Variance explained by all SNPs}

\item{S}{Selection coefficient on trait. Default = 0}

\item{Pi}{Proportion of variants that have an effect - sampled randomly. Default=1}
}
\value{
data frame
}
\description{
Generate SNP effects given MAF, h2 and selection
}
