% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theoretical_gwas.R
\name{generate_gwas_ss}
\alias{generate_gwas_ss}
\title{Create a GWAS summary dataset}
\usage{
generate_gwas_ss(beta, af, nid, vy = 1, minmaf = 0.01)
}
\arguments{
\item{beta}{Array of true beta values}

\item{af}{Array of effect allele frequency values}

\item{nid}{sample size}

\item{vy}{Variance of trait}

\item{minmaf}{minimum allowed maf. default=0.01 to prevent instability}
}
\value{
list of data frames
}
\description{
Create a GWAS summary dataset
}
