% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ld.r
\name{generate_ldobj}
\alias{generate_ldobj}
\title{Generate LD matrix objects from reference panel}
\usage{
generate_ldobj(
  outdir,
  bfile,
  regions,
  plink_bin = genetics.binaRies::get_plink_binary(),
  nthreads = 1
)
}
\arguments{
\item{outdir}{Directory in which to store the ldobj files}

\item{bfile}{Binary plink dataset}

\item{regions}{A data frame containing the independent regions (see \code{data(ldetect)})}

\item{plink_bin}{Plink executable. Default=genetics.binaRies::get_plink_binary()}

\item{nthreads}{How many threads. Default=1}
}
\value{
map file
}
\description{
Creates a set of ldobj files, each corresponding to a single independent LD region from the reference panel. It also generates a map file.
}
