% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.r
\name{get_effs}
\alias{get_effs}
\title{Get effs for two traits and make dat format}
\usage{
get_effs(x, y, g, xname = "X", yname = "Y")
}
\arguments{
\item{x}{Vector of exposure phenotype}

\item{y}{Vector of outcome phenotype}

\item{g}{Matrix of genotypes}

\item{xname}{xname}

\item{yname}{yname}
}
\value{
Data frame
}
\description{
Get effs for two traits and make dat format
}
