% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast_assoc.r
\name{gwas}
\alias{gwas}
\title{Perform association of many SNPs against phenotype}
\usage{
gwas(y, g, logistic = FALSE)
}
\arguments{
\item{y}{Vector of phenotypes}

\item{g}{Matrix of genotypes}

\item{logistic}{Use logistic regression (much slower)? Default=FALSE}
}
\value{
Data frame
}
\description{
Perform association of many SNPs against phenotype
}
