% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.r
\name{gwas}
\alias{gwas}
\title{Perform association of many SNPs against phenotype}
\usage{
gwas(y, g)
}
\arguments{
\item{y}{Vector of phenotypes}

\item{g}{Matrix of genotypes}
}
\value{
Data frame
}
\description{
Perform association of many SNPs against phenotype
}
