% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.r
\name{recode_dat}
\alias{recode_dat}
\title{Recode data to make every effect on x positive}
\usage{
recode_dat(dat, method = "intercept")
}
\arguments{
\item{dat}{Output from get_effs}

\item{method}{Default 'intercept'. Alternatively can specify 'simple'}
}
\value{
Data frame
}
\description{
Can use simple method or by pivoting around intercept
}
