% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theoretical_gwas.R
\name{sample_beta}
\alias{sample_beta}
\title{Sample beta values given standard errors}
\usage{
sample_beta(beta, se)
}
\arguments{
\item{beta}{array of beta values}

\item{se}{array of se values}
}
\value{
array of beta hats
}
\description{
Sample beta values given standard errors
}
