% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateGP-package.r
\docType{package}
\name{simulateGP-package}
\alias{simulateGP-package}
\alias{simulateGP}
\alias{simulategp}
\title{simulateGP: Functions for Simulating Genotype-Phenotype Relationships}
\description{
This package is a collection of utilities relating to simulating genotype phenotype maps
The simulations are largely used for Mendelian randomisation, and link up with the TwoSampleMR R package
}
\details{
\strong{Full documentation available here:} \url{https://explodecomputer.github.io/simulateGP}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://explodecomputer.github.io/simulateGP/}
  \item Report bugs at \url{https://github.com/explodecomputer/simulateGP/issues}
}

}
\author{
\strong{Maintainer}: Gibran Hemani \email{explodecomputer@gmail.com} (\href{https://orcid.org/0000-0003-0920-1055}{ORCID})

Authors:
\itemize{
  \item John Ferguson \email{john.ferguson@nuigalway.ie}
  \item Rita Rasteiro \email{rita.rasteiro@bristol.ac.uk} (\href{https://orcid.org/0000-0002-4217-3060}{ORCID})
}

}
