% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateGP-package.r
\docType{package}
\name{simulateGP-package}
\alias{simulateGP-package}
\alias{simulateGP}
\alias{simulategp}
\title{simulateGP: Functions for simulating genotype-phenotype relationships}
\description{
This package is a collection of utilities relating to simulating genotype phenotype maps
The simulations are largely used for Mendelian randomisation, and link up with the TwoSampleMR R package
}
\details{
**Full documentation available here:** [https://explodecomputer.github.io/simulateGP](https://explodecomputer.github.io/simulateGP)
}
