% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.r
\name{y_to_binary}
\alias{y_to_binary}
\title{Convert continuous trait to binary}
\usage{
y_to_binary(y, prevalence = NULL, threshold = NULL)
}
\arguments{
\item{y}{Phenotype vector}

\item{prevalence=NULL}{Disease prevalence}

\item{threshold=NULL}{Disease threshold}
}
\value{
Vector of binary trait
}
\description{
Convert continuous trait to binary
}

